Remove TableElementFinder. (#993)
* Remove TableElementFinder.

This class was pretty useless after latest cleanups and having it as
an attribute in library itself didn't make much sense. It was also
only used by TableElementKeywords, which somewhat inconsistently
handled finding table cells itself. This commit moves
TableElementFinder functionality into TableElementKeywords.
