+++ src/SeleniumLibrary/base/context.py
@@ -39,10 +39,6 @@ class ContextAware(object):
     def element_finder(self):
         return self.ctx._element_finder
 
-    @property
-    def table_element_finder(self):
-        return self.ctx._table_element_finder
-
     def find_element(self, locator, tag=None, required=True, parent=None):
         """Find element matching `locator`.
