+++ src/SeleniumLibrary/keywords/tableelement.py
@@ -109,7 +109,7 @@ class TableElementKeywords(LibraryComponent):
         See `Page Should Contain Element` for explanation about the
         ``loglevel`` argument.
         """
-        element = self.table_element_finder.find_by_column(locator, column, expected)
+        element = self._find_by_column(locator, column, expected)
         if element is None:
             self.ctx.log_source(loglevel)
             raise AssertionError("Table '%s' column %s did not contain text "
@@ -128,7 +128,7 @@ class TableElementKeywords(LibraryComponent):
         See `Page Should Contain Element` for explanation about the
         ``loglevel`` argument.
         """
-        element = self.table_element_finder.find_by_footer(locator, expected)
+        element = self._find_by_footer(locator, expected)
         if element is None:
             self.ctx.log_source(loglevel)
             raise AssertionError("Table '%s' footer did not contain text "
@@ -147,7 +147,7 @@ class TableElementKeywords(LibraryComponent):
         See `Page Should Contain Element` for explanation about the
         ``loglevel`` argument.
         """
-        element = self.table_element_finder.find_by_header(locator, expected)
+        element = self._find_by_header(locator, expected)
         if element is None:
             self.ctx.log_source(loglevel)
             raise AssertionError("Table '%s' header did not contain text "
@@ -171,7 +171,7 @@ class TableElementKeywords(LibraryComponent):
         See `Page Should Contain Element` for explanation about the
         ``loglevel`` argument.
         """
-        element = self.table_element_finder.find_by_row(locator, row, expected)
+        element = self._find_by_row(locator, row, expected)
         if element is None:
             self.ctx.log_source(loglevel)
             raise AssertionError("Table '%s' row %s did not contain text "
@@ -187,8 +187,47 @@ class TableElementKeywords(LibraryComponent):
         See `Page Should Contain Element` for explanation about the
         ``loglevel`` argument.
         """
-        element = self.table_element_finder.find_by_content(locator, expected)
+        element = self._find_by_content(locator, expected)
         if element is None:
             self.ctx.log_source(loglevel)
             raise AssertionError("Table '%s' did not contain text '%s'."
                                  % (locator, expected))
+
+    def _find_by_content(self, table_locator, content):
+        return self._find(table_locator, '//*', content)
+
+    def _find_by_header(self, table_locator, content):
+        return self._find(table_locator, '//th', content)
+
+    def _find_by_footer(self, table_locator, content):
+        return self._find(table_locator, '//tfoot//td', content)
+
+    def _find_by_row(self, table_locator, row, content):
+        position = self._index_to_position(row)
+        locator = '//tr[{}]'.format(position)
+        return self._find(table_locator, locator, content)
+
+    def _find_by_column(self, table_locator, col, content):
+        position = self._index_to_position(col)
+        locator = '//tr//*[self::td or self::th][{}]'.format(position)
+        return self._find(table_locator, locator, content)
+
+    def _index_to_position(self, index):
+        index = int(index)
+        if index == 0:
+            raise ValueError('Row and column indexes must be non-zero.')
+        if index > 0:
+            return str(index)
+        if index == -1:
+            return 'position()=last()'
+        return 'position()=last()-{}'.format(abs(index) - 1)
+
+    def _find(self, table_locator, locator, content):
+        table = self.find_element(table_locator)
+        elements = self.find_elements(locator, parent=table)
+        for element in elements:
+            if content is None:
+                return element
+            if element.text and content in element.text:
+                return element
+        return None
