+++ src/SeleniumLibrary/locators/__init__.py
@@ -16,5 +16,4 @@
 
 from .customlocator import CustomLocator
 from .elementfinder import ElementFinder
-from .tableelementfinder import TableElementFinder
 from .windowmanager import WindowManager
